refactor : 코드 수정
 - Connection con = null Optional 처리
