+++ src/main/java/chess/dao/PieceDao.java
@@ -1,16 +1,14 @@
-package chess.domain.piece;
+package chess.dao;
 
+import chess.domain.piece.Piece;
 import chess.domain.piece.position.Position;
 
 import java.sql.*;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 public class PieceDao {
     public Connection getConnection() {
-        Connection con = null;
+        Optional<Connection> con = null;
         String server = "localhost:13306";
         String database = "db_name";
         String option = "?useSSL=false&serverTimezone=UTC";
@@ -25,22 +23,16 @@ public class PieceDao {
         }
 
         try {
-            con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password);
+            con = Optional
+                    .of(DriverManager
+                            .getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password));
             System.out.println("정상적으로 연결되었습니다.");
         } catch (SQLException e) {
             System.err.println("연결 오류:" + e.getMessage());
             e.printStackTrace();
         }
-        return con;
-    }
 
-    public void closeConnection(Connection con) {
-        try {
-            if (con != null)
-                con.close();
-        } catch (SQLException e) {
-            System.err.println("con 오류:" + e.getMessage());
-        }
+        return con.orElseThrow(NullPointerException::new);
     }
 
     public void addPiece(Piece piece) throws SQLException {
